% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equalden.HDpackage.R
\docType{package}
\name{Equalden.HD-package}
\alias{Equalden.HD}
\alias{Equalden.HD-package}
\title{Package ‘Equalden.HD’}
\value{
\itemize{
\item{Equalden.test.HD: Performs the k-sample test proposed in Zhan and Hart (2012) for the setting of low sample size, large
dimension and independent samples, and its adaptions to dependent samples proposed in Cousido-Rocha
et. al (2018).}
}
}
\description{
This package implements three different k-sample tests in the setting of low sample size and large
dimension. The first one, proposed in Zhan and Hart (2012), test the null hypothesis that all the k-
samples come from a single distribution when the number of samples is large, the sample size is small
and the samples are independent of each other. The other tests proposed in Cousido-Rocha et. al (2018)
are an adaptation of the test of Zhan and Hart (2012) to the setting of dependent sample sizes. The type
of dependence considered in such tests is absolutely regular and strongly mixing (see Doukhan, 1995).
The standarized version of the test statistic and the p-values are computed among other things.
}
\details{
Documentation for package ‘Equalden.HD’ version 1.0


\itemize{
\item{Package: Equalden.HD}
\item{Version: 1.0}
\item{Maintainer: Marta Cousido Rocha \email{martacousido@uvigo.es}}
\item{License: GPL-2}
}
}
\references{
\itemize{
\item{Cousido-Rocha, M., de Uña-Álvarez, J. (2018). Testing equality of a large number of densities under mixing conditions. Test (preprinted)}
\item{Doukhan, P. (1995) Mixing: Properties and Examples. Springer-Verlag, New York.}
\item{Zhan, D., Hart, J. (2012) Testing equality of a large number of densities. Biometrika, 99, 1-17.}
}
}
\author{
\itemize{
\item{Cousido Rocha, Marta.}
\item{Soage González, José Carlos.}
\item{de Uña-Álvarez, Jacobo.}
}
}
